+++ druid-shell/src/backend/x11/application.rs
@@ -904,7 +904,7 @@ fn poll_with_timeout(
                 break;
             } else {
                 let millis = c_int::try_from(deadline.duration_since(now).as_millis())
-                    .unwrap_or(c_int::max_value() - 1);
+                    .unwrap_or(c_int::MAX - 1);
                 // The above .as_millis() rounds down. This means we would wake up before the
                 // deadline is reached. Add one to 'simulate' rounding up instead.
                 millis + 1