+++ druid-shell/src/lib.rs
@@ -13,7 +13,7 @@
 //! variables. Here is a list of environment variables that `druid-shell` supports:
 //!
 //! - `DRUID_SHELL_DISABLE_X11_PRESENT`: if this is set and `druid-shell` is using the `x11`
-//! backend, it will avoid using the Present extension.
+//!   backend, it will avoid using the Present extension.
 
 #![warn(rustdoc::broken_intra_doc_links)]
 #![allow(clippy::new_without_default)]