+++ druid/src/widget/split.rs
@@ -420,7 +420,7 @@ impl<T: Data> Widget<T> for Split<T> {
         self.split_point_effective = {
             let (min_limit, max_limit) = self.split_side_limits(reduced_size);
             let reduced_axis_size = self.split_axis.major(reduced_size);
-            if reduced_axis_size.is_infinite() || reduced_axis_size <= std::f64::EPSILON {
+            if reduced_axis_size.is_infinite() || reduced_axis_size <= f64::EPSILON {
                 0.5
             } else {
                 self.split_point_chosen